+++ lib/lang/de.dart
@@ -182,4 +182,6 @@ const Map<String, String> de = {
       'Verwenden Sie die native Kamera-App, um Videos aufzunehmen, anstatt die integrierte Kamera. Dies kann Probleme beim Aufzeichnen für Geräte mit geringeren Leistungsmerkmalen lösen. Diese Option funktioniert nicht auf Android 13.',
   'movies': 'Filme',
   'myMovies': 'Meine filme',
+  'noMoviesFound': 'Keine Filme gefunden',
+  'play': 'Abspielen',
 };